Add functionality to migrate single groups and single migrations
if you run a single migration, it will be ran on all groups which have
a migration with that name. You can use this to run a single migration
over all groups by just copying files with the same name.

If you don't want that to happen, just use the "new" command to create
migration files with unique names.