👌 Add a `--silent` option to the `migrate` command
This makes it so that the migrate command doesn't output anything if
there aren't any migrations being run. This is useful in combination
with the dry-run functionality: if you want to write a script that
checks if migrations need to be run, you can use dry-run in combination
with `--silent` and check for empty output. If the output is empty, no
migrations need to run.

/cc @BenSeDev